requirements.txt
mcculw
vispy
phconvert
ruamel.yaml
PyQt5
pyserial